pyproject.toml
[build-system]
requires = ["setuptools>=77"]
build-backend = "setuptools.build_meta"

[project]
name = "a2a-backend-kit"
version = "0.1.0"
description = "Shared helpers for Threads A2A backend adapters"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=1.0,<1.1",
    "fastapi>=0.115",
    "httpx>=0.27",
    "sse-starlette>=3.0",
    "pyyaml>=6.0",
    "opentelemetry-sdk>=1.27",
    "opentelemetry-exporter-otlp-proto-http>=1.27",
]

[tool.setuptools.packages.find]
where = ["src"]
